Module egli::egl

C-style lower level EGL wrapper with unsafety removed.

Use these only if higher-level EGLI abstration is not enough, and you need to manage resource clanup manually.

Constants

EGL_ALPHA_FORMAT

EGL_ALPHA_FORMAT_NONPRE

EGL_ALPHA_FORMAT_PRE

EGL_ALPHA_MASK_SIZE

EGL_ALPHA_SIZE

EGL_BACK_BUFFER

EGL_BAD_ACCESS

EGL_BAD_ALLOC

EGL_BAD_ATTRIBUTE

EGL_BAD_CONFIG

EGL_BAD_CONTEXT

EGL_BAD_CURRENT_SURFACE

EGL_BAD_DISPLAY

EGL_BAD_MATCH

EGL_BAD_NATIVE_PIXMAP

EGL_BAD_NATIVE_WINDOW

EGL_BAD_PARAMETER

EGL_BAD_SURFACE

EGL_BIND_TO_TEXTURE_RGB

EGL_BIND_TO_TEXTURE_RGBA

EGL_BLUE_SIZE

EGL_BUFFER_DESTROYED

EGL_BUFFER_PRESERVED

EGL_BUFFER_SIZE

EGL_CLIENT_APIS

EGL_COLORSPACE

EGL_COLORSPACE_LINEAR

EGL_COLORSPACE_sRGB

EGL_COLOR_BUFFER_TYPE

EGL_CONFIG_CAVEAT

EGL_CONFIG_ID

EGL_CONFORMANT

EGL_CONTEXT_CLIENT_TYPE

EGL_CONTEXT_CLIENT_VERSION

EGL_CONTEXT_LOST

EGL_CORE_NATIVE_ENGINE

EGL_DEFAULT_DISPLAY

EGL_DEPTH_SIZE

EGL_DISPLAY_SCALING

EGL_DONT_CARE

EGL_DRAW

EGL_EXTENSIONS

EGL_FALSE

EGL_GREEN_SIZE

EGL_HEIGHT

EGL_HORIZONTAL_RESOLUTION

EGL_LARGEST_PBUFFER

EGL_LEVEL

EGL_LUMINANCE_BUFFER

EGL_LUMINANCE_SIZE

EGL_MATCH_NATIVE_PIXMAP

EGL_MAX_PBUFFER_HEIGHT

EGL_MAX_PBUFFER_PIXELS

EGL_MAX_PBUFFER_WIDTH

EGL_MAX_SWAP_INTERVAL

EGL_MIN_SWAP_INTERVAL

EGL_MIPMAP_LEVEL

EGL_MIPMAP_TEXTURE

EGL_MULTISAMPLE_RESOLVE

EGL_MULTISAMPLE_RESOLVE_BOX

EGL_MULTISAMPLE_RESOLVE_DEFAULT

EGL_NATIVE_RENDERABLE

EGL_NATIVE_VISUAL_ID

EGL_NATIVE_VISUAL_TYPE

EGL_NONE

EGL_NON_CONFORMANT_CONFIG

EGL_NOT_INITIALIZED

EGL_NO_CONTEXT

EGL_NO_DISPLAY

EGL_NO_SURFACE

EGL_NO_TEXTURE

EGL_OPENGL_API

EGL_OPENGL_ES_API

EGL_OPENVG_API

EGL_OPENVG_IMAGE

EGL_PIXEL_ASPECT_RATIO

EGL_READ

EGL_RED_SIZE

EGL_RENDERABLE_TYPE

EGL_RENDER_BUFFER

EGL_RGB_BUFFER

EGL_SAMPLES

EGL_SAMPLE_BUFFERS

EGL_SINGLE_BUFFER

EGL_SLOW_CONFIG

EGL_STENCIL_SIZE

EGL_SUCCESS

EGL_SURFACE_TYPE

EGL_SWAP_BEHAVIOR

EGL_TEXTURE_2D

EGL_TEXTURE_FORMAT

EGL_TEXTURE_RGB

EGL_TEXTURE_RGBA

EGL_TEXTURE_TARGET

EGL_TRANSPARENT_BLUE_VALUE

EGL_TRANSPARENT_GREEN_VALUE

EGL_TRANSPARENT_RED_VALUE

EGL_TRANSPARENT_RGB

EGL_TRANSPARENT_TYPE

EGL_TRUE

EGL_UNKNOWN

EGL_VENDOR

EGL_VERSION

EGL_VERTICAL_RESOLUTION

EGL_VG_ALPHA_FORMAT

EGL_VG_ALPHA_FORMAT_NONPRE

EGL_VG_ALPHA_FORMAT_PRE

EGL_VG_COLORSPACE

EGL_VG_COLORSPACE_LINEAR

EGL_VG_COLORSPACE_sRGB

EGL_WIDTH

Functions

bind_api
[EGL 1.2] Set the current rendering API.

bind_tex_image
[EGL 1.1] Defines a two-dimensional texture image.

copy_buffers
[EGL 1.0] Copy EGL surface color buffer to a native pixmap.

create_context
[EGL 1.0] Create a new EGL rendering context.

create_context_with_attribs
[EGL 1.0] Create a new EGL rendering context.

create_pbuffer_from_client_buffer
[EGL 1.2] Create a new EGL pixel buffer surface bound to an OpenVG image.

create_pbuffer_surface
[EGL 1.0] Create a new EGL pixel buffer surface.

create_pixmap_surface
[EGL 1.0] Create a new EGL pixmap surface.

create_window_surface
[EGL 1.0] Create a new EGL window surface.

create_window_surface_with_attribs
[EGL 1.0] Create a new EGL window surface.

destroy_context
[EGL 1.0] Destroy an EGL rendering context.

destroy_surface
[EGL 1.0] Destroy an EGL surface.

get_config_attrib
[EGL 1.0] Return information about an EGL frame buffer configuration.

get_configs
[EGL 1.0] Return a list of all EGL frame buffer configurations for a display.

get_current_context
[EGL 1.4] Return the current EGL rendering context.

get_current_display
[EGL 1.0] Return the display for the current EGL rendering context.

get_current_surface
[EGL 1.0] Return the read or draw surface for the current EGL rendering context.

get_display
[EGL 1.0] Return an EGL display connection.

get_error
[EGL 1.0] Return error information.

get_filtered_configs
[EGL 1.0] Return a list of EGL frame buffer configurations that match specified attributes.

get_proc_address
[EGL 1.0] Return a GL or an EGL extension function.

initialize
[EGL 1.0] Initialize an EGL display connection.

initialize_and_get_version
[EGL 1.0] Initialize an EGL display connection and get EGL version.

make_current
[EGL 1.0] Attach an EGL rendering context to EGL surfaces.

num_configs
[EGL 1.0] Return the total number of all available display configs.

num_filtered_configs
[EGL 1.0] Return the total number of display configs that match specified attributes.

query_api
[EGL 1.2] Query the current rendering API.

query_context
[EGL 1.0] Return EGL rendering context information.

query_string
[EGL 1.0] Return a string describing an EGL display connection.

query_surface
[EGL 1.0] Return EGL surface information.

release_tex_image
[EGL 1.1] Releases a color buffer that is being used as a texture.

release_thread
[EGL 1.2] Release EGL per-thread state.

surface_attrib
[EGL 1.1] Set an EGL surface attribute.

swap_buffers
[EGL 1.0] Post EGL surface color buffer to a native window.

swap_interval
[EGL 1.1] Specifies the minimum number of video frame periods per buffer swap for the window associated with the current context.

terminate
[EGL 1.0] Terminate an EGL display connection.

wait_client
[EGL 1.2] Complete client API execution prior to subsequent native rendering calls.

wait_gl
[EGL 1.0] Complete GL execution prior to subsequent native rendering calls.

wait_native
[EGL 1.0] Complete native execution prior to subsequent GL rendering calls.

Type Definitions

EGLBoolean

EGLClientBuffer

EGLConfig

EGLContext

EGLDisplay

EGLNativeDisplayType

EGLNativePixmapType

EGLNativeWindowType

EGLSurface

EGLenum

EGLint