Struct egli::FrameBufferConfigRef

pub struct FrameBufferConfigRef { /* fields omitted */ }

[EGL 1.0] Reference to frame buffer configuration.

This is not a RAII object, so nothing will be cleaned up when it is dropped.

Instead, the individual methods of this object may return errors if they are used after the display is terminated.

Methods

impl FrameBufferConfigRef

fn from_native(display_id: EGLDisplay,
config_handle: EGLConfig)
-> FrameBufferConfigRef

fn handle(&self) -> EGLConfig

Get native config handle.

fn alpha_size(&self) -> Result<u32>

Returns the number of bits of alpha stored in the color buffer.

Calls eglGetConfigAttrib with EGL_ALPHA_SIZE attribute.

fn alpha_mask_size(&self) -> Result<u32>

Returns the number of bits in the alpha mask buffer.

Calls eglGetConfigAttrib with EGL_ALPHA_MASK_SIZE attribute.

fn bind_to_texture_rgb(&self) -> Result<bool>

Returns whether color buffers can be bound to an RGB texture.

Calls eglGetConfigAttrib with EGL_BIND_TO_TEXTURE_RGB attribute.

fn bind_to_texture_rgba(&self) -> Result<bool>

Returns whether buffers can be bound to an RGBA texture.

Calls eglGetConfigAttrib with EGL_BIND_TO_TEXTURE_RGBA attribute.

fn blue_size(&self) -> Result<u32>

Returns the number of bits of blue stored in the color buffer.

Calls eglGetConfigAttrib with EGL_BLUE_SIZE attribute.

fn buffer_size(&self) -> Result<u32>

Returns the depth of the color buffer. It is the sum of EGL_RED_SIZE, EGL_GREEN_SIZE, EGL_BLUE_SIZE, and EGL_ALPHA_SIZE.

Calls eglGetConfigAttrib with EGL_BUFFER_SIZE attribute.

fn color_buffer_type(&self) -> Result<ColorBufferType>

Returns the color buffer type. Possible types are EGL_RGB_BUFFER and EGL_LUMINANCE_BUFFER.

Calls eglGetConfigAttrib with EGL_COLOR_BUFFER_TYPE attribute.

fn config_caveat(&self) -> Result<ConfigCaveat>

Returns the caveats for the frame buffer configuration. Possible caveat values are EGL_NONE, EGL_SLOW_CONFIG, and EGL_NON_CONFORMANT.

Calls eglGetConfigAttrib with EGL_CONFIG_CAVEAT attribute.

fn config_id(&self) -> Result<i32>

Returns the ID of the frame buffer configuration.

Calls eglGetConfigAttrib with EGL_CONFIG_ID attribute.

fn conformant(&self) -> Result<Type>

Returns a bitmask indicating which client API contexts created with respect to this config are conformant.

EGL_CONFORMANT is supported only if the EGL version is 1.3 or greater.

Calls eglGetConfigAttrib with EGL_CONFORMANT attribute.

fn depth_size(&self) -> Result<u32>

Returns the number of bits in the depth buffer.

Calls eglGetConfigAttrib with EGL_DEPTH_SIZE attribute.

fn green_size(&self) -> Result<u32>

Returns the number of bits of green stored in the color buffer.

Calls eglGetConfigAttrib with EGL_GREEN_SIZE attribute.

fn level(&self) -> Result<i32>

Returns the frame buffer level. Level zero is the default frame buffer. Positive levels correspond to frame buffers that overlay the default buffer and negative levels correspond to frame buffers that underlay the default buffer.

Calls eglGetConfigAttrib with EGL_LEVEL attribute.

fn luminance_size(&self) -> Result<u32>

Returns the number of bits of luminance stored in the luminance buffer.

Calls eglGetConfigAttrib with EGL_LUMINANCE_SIZE attribute.

fn max_pbuffer_width(&self) -> Result<i32>

Returns the maximum width of a pixel buffer surface in pixels.

Calls eglGetConfigAttrib with EGL_MAX_PBUFFER_WIDTH attribute.

fn max_pbuffer_height(&self) -> Result<i32>

Returns the maximum height of a pixel buffer surface in pixels.

Calls eglGetConfigAttrib with EGL_MAX_PBUFFER_HEIGHT attribute.

fn max_pbuffer_pixels(&self) -> Result<i32>

Returns the maximum size of a pixel buffer surface in pixels.

Calls eglGetConfigAttrib with EGL_MAX_PBUFFER_PIXELS attribute.

fn max_swap_interval(&self) -> Result<i32>

Returns the maximum value that can be passed to eglSwapInterval.

Calls eglGetConfigAttrib with EGL_MAX_SWAP_INTERVAL attribute.

fn min_swap_interval(&self) -> Result<i32>

Returns the minimum value that can be passed to eglSwapInterval.

Calls eglGetConfigAttrib with EGL_MIN_SWAP_INTERVAL attribute.

fn native_renderable(&self) -> Result<bool>

Returns whether native rendering APIs can render into the surface.

Calls eglGetConfigAttrib with EGL_NATIVE_RENDERABLE attribute.

fn native_visual_id(&self) -> Result<i32>

Returns the ID of the associated native visual.

Calls eglGetConfigAttrib with EGL_NATIVE_VISUAL_ID attribute.

fn native_visual_type(&self) -> Result<i32>

Returns the type of the associated native visual.

Calls eglGetConfigAttrib with EGL_NATIVE_VISUAL_TYPE attribute.

fn red_size(&self) -> Result<u32>

Returns the number of bits of red stored in the color buffer.

Calls eglGetConfigAttrib with EGL_RED_SIZE attribute.

fn renderable_type(&self) -> Result<Type>

Returns a bitmask indicating the types of supported client API contexts.

Calls eglGetConfigAttrib with EGL_RENDERABLE_TYPE attribute.

fn sample_buffers(&self) -> Result<i32>

Returns the number of multisample buffers.

Calls eglGetConfigAttrib with EGL_SAMPLE_BUFFERS attribute.

fn samples(&self) -> Result<i32>

Returns the number of samples per pixel.

Calls eglGetConfigAttrib with EGL_SAMPLES attribute.

fn stencil_size(&self) -> Result<u32>

Returns the number of bits in the stencil buffer.

Calls eglGetConfigAttrib with EGL_STENCIL_SIZE attribute.

fn surface_type(&self) -> Result<Type>

Returns a bitmask indicating the types of supported EGL surfaces.

Calls eglGetConfigAttrib with EGL_SURFACE_TYPE attribute.

fn transparent_type(&self) -> Result<TransparentType>

Returns the type of supported transparency. Possible transparency values are: EGL_NONE, and EGL_TRANSPARENT_RGB.

Calls eglGetConfigAttrib with EGL_TRANSPARENT_TYPE attribute.

fn transparent_red_value(&self) -> Result<u32>

Returns the transparent red value.

Calls eglGetConfigAttrib with EGL_TRANSPARENT_RED_VALUE attribute.

fn transparent_green_value(&self) -> Result<u32>

Returns the transparent green value.

Calls eglGetConfigAttrib with EGL_TRANSPARENT_GREEN_VALUE attribute.

fn transparent_blue_value(&self) -> Result<u32>

Returns the transparent blue value.

Calls eglGetConfigAttrib with EGL_TRANSPARENT_BLUE_VALUE attribute.